feat: Added new functionality to update a movie by its id and implemented "Service Interface with Service Implementation" to improve dependency injection and decoupling
