+++ src/main/java/com/br/movies/domain/film/payloads/MovieResponse.java
@@ -1,5 +1,6 @@
 package com.br.movies.domain.film.payloads;
 
+import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -17,6 +18,8 @@ public class MovieResponse {
     private String description;
     private List<String> categories;
     private String ageRating;
+
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy")
     private Date year;
     private String image;
 }
